mcp_server/pyproject.toml
[project]
name = "mcp_server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=0.4.0",
  "snowflake-connector-python>=3.10.0",
  "kafka-python>=2.0.2",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0"
]

[project.scripts]
mcp-server = "mcp_server.server:main"
